move backend functions to Explorer.Backend
